KeywordSearch/test/unit/src/org/sleuthkit/autopsy/keywordsearch/QueryParserTest.java
/*
 * Autopsy Forensic Browser
 *
 * Copyright 2011-2019 Basis Technology Corp.
 * Contact: carrier <at> sleuthkit <dot> org
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package org.sleuthkit.autopsy.keywordsearch;

import org.junit.Test;

import java.util.List;
import java.util.Map;

import static org.junit.Assert.assertEquals;

public class QueryParserTest {

  @Test
  public void getFieldTermsMap() {
    Map<String, List<String>> map = QueryParser.getFieldTermsMap("content_ja:雨 content_ja:降る");
    List<String> terms = map.get("content_ja");
    assertEquals(2, terms.size());
    assertEquals("雨", terms.get(0));
    assertEquals("降る", terms.get(1));
  }
}
